Began shop implementation and generally just cleaned up some files
