+++ Project-1/Assets/Scripts/ItemManager.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ItemManager : MonoBehaviour
+{
+    // bools for items that the destroyable objs will check against during gameplay
+    public bool extraDamage;
+    public bool extraMoney;
+    public bool extraTime;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
