+++ Project-1/Assets/Scripts/ItemManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9ca9646a34a860545a7f98789e6ce0e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
