+++ Project-1/Assets/Scripts/Manager.cs.meta
