+++ Project-1/Assets/Scripts/MoneyManager.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+//script for tracking player's money
+
+public class MoneyManager : MonoBehaviour
+{
+    public int money = 0;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
