+++ Project-1/Assets/Scripts/MoneyManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 00fa24c00b3e9584098a0b32a0ece537
+guid: d5595cde68f4dda45b9abd2a5ebaf3ff
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
